+++ build_tools/circle/build_doc.sh
@@ -6,12 +6,14 @@ set -e
 # documentation: a pull request that does not change any file in doc/ or
 # examples/ folder should be skipped unless the "[doc: build]" is found the
 # commit message.
-BUILD_DOC=`python build_tools/circle/check_build_doc.py`
-echo -e $BUILD_DOC
-if [[ $BUILD_DOC == "SKIP:"* ]]; then
-    touch ~/log.txt  # the "test" segment needs that file
-    exit 0
-fi
+# 
+# Disabling this for now. rth.
+#BUILD_DOC=`python build_tools/circle/check_build_doc.py`
+#echo -e $BUILD_DOC
+#if [[ $BUILD_DOC == "SKIP:"* ]]; then
+#    touch ~/log.txt  # the "test" segment needs that file
+#    exit 0
+#fi
 
 # Installing required system packages to support the rendering of match
 # notation in the HTML documentation