+++ examples/REST_clustering.py
@@ -24,8 +24,8 @@ dataset_name = "treclegal09_2k_subset"     # see list of available datasets
 
 BASE_URL = "http://localhost:5001/api/v0"  # FreeDiscovery server URL
 
-print(" 0. Load the test dataset")
-url = BASE_URL + '/datasets/{}'.format(dataset_name)
+print(" 0. Load the example dataset")
+url = BASE_URL + '/example-dataset/{}'.format(dataset_name)
 print(" POST", url)
 res = requests.get(url).json()